move upgrade tab logic
