+++ Assets/Scenes/MainScene.unity
@@ -2158,7 +2158,7 @@ RectTransform:
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
-  m_AnchorMax: {x: 1, y: 1}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: -0.000030518, y: 0}
   m_SizeDelta: {x: 13.35, y: 16.364}
   m_Pivot: {x: 0.5, y: 0.5}
@@ -2345,13 +2345,14 @@ GameObject:
   - component: {fileID: 1053374193}
   - component: {fileID: 1053374195}
   - component: {fileID: 1053374194}
+  - component: {fileID: 1053374196}
   m_Layer: 5
   m_Name: UpgradesTab
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &1053374193
 RectTransform:
   m_ObjectHideFlags: 0
@@ -2415,6 +2416,21 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1053374192}
   m_CullTransparentMesh: 1
+--- !u!114 &1053374196
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1053374192}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fa49aa9001c517b41acb15c767ac22bf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  buttonContainer: {fileID: 1637579001}
+  buttonPrefab: {fileID: 2413596800833516689, guid: 1ca43ff8c3d5c654d80d362b1c043069, type: 3}
+  descriptionText: {fileID: 1950059080}
 --- !u!114 &1084491240 stripped
 MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 5566100957997582797, guid: 5acddee02b59fde49aa90e4204f2ae24, type: 3}
@@ -6065,13 +6081,11 @@ MonoBehaviour:
   eventText: {fileID: 153595740}
   gameOverTab: {fileID: 565476646}
   currencyCounter: {fileID: 1237179554}
+  upgradeTab: {fileID: 0}
   changeWindowButtonObj: {fileID: 1532701914}
   saveInfoText: {fileID: 2055665938}
   volumeSlider: {fileID: 44669676}
   volumeValueText: {fileID: 778063330}
-  upgradeButtonContainer: {fileID: 1637579001}
-  upgradeButtonPrefab: {fileID: 2413596800833516689, guid: 1ca43ff8c3d5c654d80d362b1c043069, type: 3}
-  upgradeDescriptionText: {fileID: 1950059080}
   energyCounter: {fileID: 1880873462}
   energySafeColor: {r: 0.93725497, g: 0.7411765, b: 0.41176474, a: 1}
   energyDangerColor: {r: 0.16862746, g: 0.27058825, b: 0.30980393, a: 1}
